refactor(language-service): split tests into multiple targets
The helps with fighting flakiness
