+++ packages/language-service/test/grp1/diagnostic_spec.ts
@@ -10,7 +10,7 @@ import {ErrorCode, ngErrorCode} from '@angular/compiler-cli/src/ngtsc/diagnostic
 import {initMockFileSystem} from '@angular/compiler-cli/src/ngtsc/file_system/testing';
 import ts from 'typescript';
 
-import {createModuleAndProjectWithDeclarations, LanguageServiceTestEnv} from '../testing';
+import {createModuleAndProjectWithDeclarations, LanguageServiceTestEnv} from '../../testing';
 
 describe('getSemanticDiagnostics', () => {
   let env: LanguageServiceTestEnv;
