+++ packages/language-service/test/grp1/get_outlining_spans_spec.ts
@@ -9,7 +9,7 @@
 import {initMockFileSystem} from '@angular/compiler-cli/src/ngtsc/file_system/testing';
 import ts from 'typescript';
 
-import {createModuleAndProjectWithDeclarations, LanguageServiceTestEnv} from '../testing';
+import {createModuleAndProjectWithDeclarations, LanguageServiceTestEnv} from '../../testing';
 
 describe('get outlining spans', () => {
   beforeEach(() => {
