+++ packages/language-service/test/grp1/get_template_location_for_component_spec.ts
@@ -13,7 +13,7 @@ import {
   createModuleAndProjectWithDeclarations,
   humanizeDocumentSpanLike,
   LanguageServiceTestEnv,
-} from '../testing';
+} from '../../testing';
 
 describe('get template location for component', () => {
   beforeEach(() => {
