+++ packages/language-service/test/grp1/references_and_rename_spec.ts
@@ -17,7 +17,7 @@ import {
   LanguageServiceTestEnv,
   OpenBuffer,
   Project,
-} from '../testing';
+} from '../../testing';
 
 describe('find references and rename locations', () => {
   let env: LanguageServiceTestEnv;
