+++ packages/language-service/test/grp1/semantic_tokens_spec.ts
@@ -9,8 +9,8 @@
 import ts from 'typescript';
 
 import {initMockFileSystem} from '@angular/compiler-cli/src/ngtsc/file_system/testing';
-import {LanguageServiceTestEnv, OpenBuffer} from '../testing';
-import {TokenEncodingConsts, TokenType, TokenModifier} from '../src/semantic_tokens';
+import {TokenEncodingConsts, TokenModifier, TokenType} from '../../src/semantic_tokens';
+import {LanguageServiceTestEnv, OpenBuffer} from '../../testing';
 
 describe('semantic tokens', () => {
   beforeEach(() => {
