+++ packages/language-service/test/grp1/signal_queries_refactoring_action_spec.ts
@@ -8,7 +8,7 @@
 
 import {initMockFileSystem} from '@angular/compiler-cli/src/ngtsc/file_system/testing';
 
-import {createModuleAndProjectWithDeclarations, LanguageServiceTestEnv} from '../testing';
+import {createModuleAndProjectWithDeclarations, LanguageServiceTestEnv} from '../../testing';
 
 describe('Signal queries refactoring action', () => {
   let env: LanguageServiceTestEnv;
