+++ packages/language-service/test/grp4/signature_help_spec.ts
@@ -7,9 +7,9 @@
  */
 
 import {initMockFileSystem} from '@angular/compiler-cli/src/ngtsc/file_system/testing';
-import {getText} from '../testing/src/util';
+import {getText} from '../../testing/src/util';
 
-import {LanguageServiceTestEnv, OpenBuffer} from '../testing';
+import {LanguageServiceTestEnv, OpenBuffer} from '../../testing';
 
 describe('signature help', () => {
   beforeEach(() => {
